fix: update command signature and migration data types

+++ src/Database/migrations/2025_11_21_203912_create_table_midtrans_transactions.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('midtrans_transactions', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('order_id')->index();
+            $table->unsignedInteger('order_id')->index();
             $table->string('midtrans_order_id')->unique()->index();
             $table->string('payment_type')->nullable();
             $table->string('bank')->nullable();
